Fixed a bunch of errors and almost lost my mind

+++ src/AnimatedLEDStripConcurrent.kt
@@ -10,14 +10,14 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
         for (i in 0 until numLEDs) shuffleArray.add(i)
     }
 
-    fun alternate(colorValues1: ColorContainer, colorValues2: ColorContainer, delay: Int = 1000, delayMod: Double = 1) {
+    fun alternate(colorValues1: ColorContainer, colorValues2: ColorContainer, delay: Int = 1000, delayMod: Double = 1.0) {
         setStripColor(colorValues1)
-        delay(delay * delayMod)
+        delay(delay * delayMod.toInt())
         setStripColor(colorValues2)
-        delay(delay * delayMod)
+        delay(delay * delayMod.toInt())
     }
 
-    fun alternate(r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 1000, delayMod: Double = 1) = alternate(ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In), delay, delayMod)
+    fun alternate(r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 1000, delayMod: Double = 1.0) = alternate(ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In), delay, delayMod)
 
     fun fadePixelRed(pixel: Int, startIntensity: Int, endIntensity: Int, revertAtCompletion: Boolean) {
         val fadeDirection: Int = if (startIntensity > endIntensity) -1 else if (startIntensity < endIntensity) 1 else 0
@@ -96,13 +96,13 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
         //Todo: fill in function
     }
 
-    fun multiPixelRun(spacing: Int, chaseDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 100, delayMod: Double = 1) {
+    fun multiPixelRun(spacing: Int, chaseDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 100, delayMod: Double = 1.0) {
         if (chaseDirection == Direction.BACKWARD) {
             for (q in 0 until spacing) {
                 setStripColor(colorValues2)
                 for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues1)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
                 for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues2)
             }
         } else if (chaseDirection == Direction.FORWARD) {
@@ -110,21 +110,21 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
                 setStripColor(colorValues2)
                 for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues1)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
                 for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues2)
             }
         }
     }
 
-    fun multiPixelRun(spacing: Int, chaseDirection: Direction, r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 100, delayMod: Double = 1) = multiPixelRun(spacing, chaseDirection, ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In), delay, delayMod)
+    fun multiPixelRun(spacing: Int, chaseDirection: Direction, r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 100, delayMod: Double = 1.0) = multiPixelRun(spacing, chaseDirection, ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In), delay, delayMod)
 
-    fun multiPixelRunToColor(spacing: Int, chaseDirection: Direction, colorValues1: ColorContainer, delay: Int = 150, delayMod: Double = 1) {
+    fun multiPixelRunToColor(spacing: Int, chaseDirection: Direction, colorValues1: ColorContainer, delay: Int = 150, delayMod: Double = 1.0) {
         if (chaseDirection == Direction.BACKWARD) {
             for (q in 0 until spacing) {
 //                setStripColor(colorValues2)
                 for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues1)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
 //                for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues2)
             }
         } else if (chaseDirection == Direction.FORWARD) {
@@ -132,43 +132,43 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
 //                setStripColor(colorValues2)
                 for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues1)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
 //                for (i in 0 until ledStrip.numPixels - 1 step spacing) setPixelColor(i + (-(q - (spacing - 1))), colorValues2)
             }
         }
     }
 
-    fun multiPixelRunToColor(spacing: Int, chaseDirection: Direction, r1In: Int, g1In: Int, b1In: Int, delay: Int = 150, delayMod: Double = 1) = multiPixelRunToColor(spacing, chaseDirection, ColorContainer(r1In, g1In, b1In), delay, delayMod)
+    fun multiPixelRunToColor(spacing: Int, chaseDirection: Direction, r1In: Int, g1In: Int, b1In: Int, delay: Int = 150, delayMod: Double = 1.0) = multiPixelRunToColor(spacing, chaseDirection, ColorContainer(r1In, g1In, b1In), delay, delayMod)
 
-    fun pixelRun(movementDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 50, delayMod: Double = 1) {
+    fun pixelRun(movementDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 50, delayMod: Double = 1.0) {
         setStripColor(colorValues2)
         if (movementDirection == Direction.FORWARD) {
             for (q in 0 until ledStrip.numPixels) {
                 setPixelColor(q, colorValues1)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
                 setPixelColor(q, colorValues2)
             }
         } else if (movementDirection == Direction.BACKWARD) {
             for (q in ledStrip.numPixels - 1 downTo 0) {
                 setPixelColor(q, colorValues1)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
                 setPixelColor(q, colorValues2)
             }
         }
     }
 
-    fun pixelRun(movementDirection: Direction, r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 50, delayMod: Double = 1) = pixelRun(movementDirection, ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In), delay, delayMod)
+    fun pixelRun(movementDirection: Direction, r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 50, delayMod: Double = 1.0) = pixelRun(movementDirection, ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In), delay, delayMod)
 
-    fun pixelRunWithTrail(movementDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 50, delayMod: Double = 1) {
+    fun pixelRunWithTrail(movementDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 50, delayMod: Double = 1.0) {
         if (movementDirection == Direction.FORWARD) {
             for (q in 0 until ledStrip.numPixels) {
                 for (i in 0 until ledStrip.numPixels - 1) {
                     setPixelColor(i, blend(getPixelColor(i), colorValues2, 60))
                 }
                 setPixelColor(q, colorValues1)
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
             }
         } else if (movementDirection == Direction.BACKWARD) {
             for (q in ledStrip.numPixels - 1 downTo 0) {
@@ -176,14 +176,14 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
                     setPixelColor(i, blend(getPixelColor(i), colorValues2, 60))
                 }
                 setPixelColor(q, colorValues1)
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
             }
         }
     }
 
-    fun pixelRunWithTrail(movementDirection: Direction, r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 50, delayMod: Double = 1) = pixelRunWithTrail(movementDirection, ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In, delay, delayMod))
+    fun pixelRunWithTrail(movementDirection: Direction, r1In: Int, g1In: Int, b1In: Int, r2In: Int, g2In: Int, b2In: Int, delay: Int = 50, delayMod: Double = 1.0) = pixelRunWithTrail(movementDirection, ColorContainer(r1In, g1In, b1In), ColorContainer(r2In, g2In, b2In), delay, delayMod)
 
-    fun smoothChase(palette: RGBPalette16, movementDirection: Direction, brightness: Int = 255, delay: Int = 50, delayMod: Double = 1) {
+    fun smoothChase(palette: RGBPalette16, movementDirection: Direction, brightness: Int = 255, delay: Int = 50, delayMod: Double = 1.0) {
         for (i in 0 until numLEDs) {
             colorListFromPalette(palette, i)
         }
@@ -204,28 +204,28 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
 
     }
 
-    fun sparkle(sparkleColor: ColorContainer, delay: Int = 50, delayMod: Double = 1) {
+    fun sparkle(sparkleColor: ColorContainer, delay: Int = 50, delayMod: Double = 1.0) {
         var originalColor: ColorContainer
         shuffleArray.shuffle()
         for (i in 0 until ledStrip.numPixels) {
             originalColor = getPixelColor(shuffleArray[i])
             setPixelColor(shuffleArray[i], sparkleColor)
             show()
-            delay(delay * delayMod)
+            delay(delay * delayMod.toInt())
             setPixelColor(shuffleArray[i], originalColor)
         }
     }
 
-    fun sparkle(rIn: Int, gIn: Int, bIn: Int, delay: Int = 50, delayMod: Double = 1) = sparkle(ColorContainer(rIn, gIn, bIn), delay, delayMod)
+    fun sparkle(rIn: Int, gIn: Int, bIn: Int, delay: Int = 50, delayMod: Double = 1.0) = sparkle(ColorContainer(rIn, gIn, bIn), delay, delayMod)
 
-    fun sparkleCC(sparkleColor: ColorContainer, delay: Int = 50, delayMod: Double = 1) {
+    fun sparkleCC(sparkleColor: ColorContainer, delay: Int = 50, delayMod: Double = 1.0) {
         val deferred = (0 until ledStrip.numPixels).map { n ->
             GlobalScope.async {
                 val originalColor: ColorContainer = getPixelColor(n)
                 delay(random().toInt() % 4950)
                 setPixelColor(n, sparkleColor)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
                 setPixelColor(n, originalColor)
             }
         }
@@ -234,27 +234,27 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
         }
     }
 
-    fun sparkleCC(rIn: Int, gIn: Int, bIn: Int, delay: Int = 50, delayMod: Double = 1) = sparkleCC(ColorContainer(rIn, gIn, bIn), delay, delayMod)
+    fun sparkleCC(rIn: Int, gIn: Int, bIn: Int, delay: Int = 50, delayMod: Double = 1.0) = sparkleCC(ColorContainer(rIn, gIn, bIn), delay, delayMod)
 
-    fun sparkleToColor(destinationColor: ColorContainer, delay: Int = 50, delayMod: Double = 1) {
+    fun sparkleToColor(destinationColor: ColorContainer, delay: Int = 50, delayMod: Double = 1.0) {
         shuffleArray.shuffle()
         for (i in 0 until ledStrip.numPixels) {
             setPixelColor(shuffleArray[i], destinationColor)
             show()
-            delay(delay * delayMod)
+            delay(delay * delayMod.toInt())
         }
     }
 
-    fun sparkleToColor(rIn: Int, gIn: Int, bIn: Int, delay: Int = 50, delayMod: Double = 1) = sparkleToColor(ColorContainer(rIn, gIn, bIn), delay, delayMod)
+    fun sparkleToColor(rIn: Int, gIn: Int, bIn: Int, delay: Int = 50, delayMod: Double = 1.0) = sparkleToColor(ColorContainer(rIn, gIn, bIn), delay, delayMod)
 
-    fun stack(stackDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 10, delayMod: Double = 1) {
+    fun stack(stackDirection: Direction, colorValues1: ColorContainer, colorValues2: ColorContainer = CCBlack, delay: Int = 10, delayMod: Double = 1.0) {
         if (stackDirection == Direction.FORWARD) {
             setStripColor(colorValues2)
             for (q in ledStrip.numPixels - 1 downTo 0) {
                 for (i in 0 until q) {
                     setPixelColor(i, colorValues1)
                     show()
-                    delay(delay * delayMod)
+                    delay(delay * delayMod.toInt())
                     setPixelColor(i, colorValues2)
                 }
                 setPixelColor(q, colorValues1)
@@ -266,7 +266,7 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
                 for (i in q - 1 downTo 0) {
                     setPixelColor(i, colorValues1)
                     show()
-                    delay(delay * delayMod)
+                    delay(delay * delayMod.toInt())
                     setPixelColor(i, colorValues2)
                 }
                 setPixelColor(q, colorValues1)
@@ -284,23 +284,23 @@ class AnimatedLEDStripConcurrent(numLEDs: Int, pin: Int) : LEDStripConcurrent(nu
         }
     }
 
-    private fun wipe(colorValues: ColorContainer, wipeDirection: Direction, delay: Int = 10, delayMod: Double = 1) {
+    private fun wipe(colorValues: ColorContainer, wipeDirection: Direction, delay: Int = 10, delayMod: Double = 1.0) {
         if (wipeDirection == Direction.BACKWARD) {
             for (i in ledStrip.numPixels - 1 downTo 0) {
                 setPixelColor(i, colorValues)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
             }
         } else if (wipeDirection == Direction.FORWARD) {
             for (i in 0 until ledStrip.numPixels) {
                 setPixelColor(i, colorValues)
                 show()
-                delay(delay * delayMod)
+                delay(delay * delayMod.toInt())
             }
         }
     }
 
-    fun wipe(rIn: Int, gIn: Int, bIn: Int, wipeDirection: Direction, delay: Int = 10, delayMod: Double = 1) = wipe(ColorContainer(rIn, gIn, bIn), wipeDirection, delay, delayMod)
+    fun wipe(rIn: Int, gIn: Int, bIn: Int, wipeDirection: Direction, delay: Int = 10, delayMod: Double = 1.0) = wipe(ColorContainer(rIn, gIn, bIn), wipeDirection, delay, delayMod)
 
 
 }
